Improve testing infrastructure
Moved most of existing tests/conftest.py to tests/integration/conftest.py

Preserved the docker setup/teardown bits

Force `repo` pytest fixture to report the correct Elasticsearch repository name from the ENV vars and manually apply that setting to the testbed object
